+++ src/tasks/brew.rs
@@ -22,8 +22,11 @@ pub fn env(mut exports: Exports) -> Exports {
                 }
             })
             .collect();
-        paths.append(&mut exports.path);
-        exports.path = paths;
+
+        // we add homebrew to the end of PATH,
+        // as we want to treat it more like a cross-distribution fallback,
+        // rather than the primary source of packages (for now, at least)
+        exports.path.append(&mut paths);
 
         // TODO: parse and export the output from `brew shellenv`
     }